+++ actix-web-lab/src/lib.rs
@@ -69,6 +69,7 @@ mod test_request_macros;
 mod test_response_macros;
 mod test_services;
 mod url_encoded_form;
+mod x_forwarded_prefix;
 
 // public API
 pub mod body;